extract MovieImage, Header, StoryCard. Add MovieDetails.
