+++ netflix-roulette/src/components/App/App.js
@@ -3,13 +3,14 @@ import React, { Suspense } from 'react';
 import ErrorBoundary from './ErrorBoundary/ErrorBoundaryContainer';
 import Footer from '../shared/footer/Footer';
 import { wordings } from '/src/locales/wordings';
+import MovieDetails from '../features/MovieDetails/MovieDetailsContainer';
 
 const Home = React.lazy(() => import('../features/Home/HomeContainer'));
 
 const App = () => (
   <Suspense fallback={wordings.loading}>
     <ErrorBoundary>
-      <Home />
+      <MovieDetails />
       <Footer />
     </ErrorBoundary>
   </Suspense>
