+++ netflix-roulette/src/components/features/home/Home.module.scss
@@ -1,30 +1,18 @@
-.movieListSection {
-  background-color: #232323;
-}
-
-.border {
-  margin-left: 64px;
-  border-bottom: solid thin #424242;
-}
-
-.moviesCount {
-  padding: 24px 64px;
-  color: #ffffff;
-  font-size: 16px;
-}
-
-.movieList {
-  padding-left: 64px;
-}
-
-.confirmButton {
-  float: right;
-  width: 150px;
-  height: 50px;
-}
-
 .submitButton {
   float: right;
   width: 155px;
   height: 55px;
+}
+
+.searchSection{
+  position: relative;
+  padding: 24px 64px;
+  z-index: 1;
+}
+
+.searchDescription {
+  text-transform: uppercase;
+  font-size: 76px;
+  margin-top: 64px;
+  margin-left: 140px;
 }
