+++ netflix-roulette/src/components/shared/MovieCard/MovieCard.module.scss
@@ -4,33 +4,6 @@
   padding-bottom: 24px;
 }
 
-.image {
-  margin-bottom: 8px;
-  height: 450px;
-  background-color: 232323;
-  width: 300px;
-}
-
-.noImageWrapper {
-  @extend .image;
-
-  border: solid thin;
-}
-
-.image img {
-  min-width: 300px;
-  width: 300px;
-  height: 450px;
-}
-
-.noImage {
-  font-size: 80px;
-  top: 30%;
-  position: absolute;
-  font-weight: 23px;
-  left: 30%;
-}
-
 .nameWraper {
   width: 100%;
   display: flex;
