+++ netflix-roulette/src/components/shared/MoviesList/data.json
@@ -186,4 +186,3 @@
     "offset": 0,
     "limit": 10
   }
-
