+++ netflix-roulette/src/components/shared/header/Header.module.scss
@@ -1,38 +1,3 @@
-.searchSectionWrapper {
-  margin-bottom: 16px;
-  position: relative;
-  height: 600px;
-  width: 100%;
-}
-
-.background {
-  z-index: 0;
-  position: absolute;
-  top: -70px;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  overflow: hidden; 
-  background-size: cover;
-  background-repeat: no-repeat;
-}
-
-.gradient {
-  position: absolute;
-  top: 0;
-  bottom: 0;
-  right: 0;
-  left: 0;
-  background: rgba(0,0,0,.4);
-  background-image: linear-gradient(to top,rgba(0,0,0,.8) 0,rgba(0,0,0,0) 60%,rgba(0,0,0,.8) 100%);
-}
-  
-.searchSection{
-  position: relative;
-  padding: 24px 64px;
-  z-index: 1;
-}
-  
 .header {
   width: 100%;
   margin: 0 auto;
@@ -41,12 +6,4 @@
   flex-direction: row;
   justify-content: space-between;
   align-items: center;
-}
-
-.searchDescription {
-  text-transform: uppercase;
-  font-size: 76px;
-  color: #ffffff;
-  margin-top: 64px;
-  margin-left: 140px;
 }
