+++ netflix-roulette/src/index.scss
@@ -6,4 +6,5 @@ body {
   margin: 0 auto;
   font-family: 'Quicksand', sans-serif;
   font-size: 16px;
+  color: #ffffff;
 }
